+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# confluent-kafka-javascript 1.x.x
+
+## Fixes
+1. Handle anyOf/allOf in JSON transforms (#479)
+
 
 # confluent-kafka-javascript 1.9.0